Fixed intermittent bug with printing where the printer doesn't always find the file we just uploaded. It looks like the /data/ prefix works more consistently than the 0:/user/ prefix.
The long term fix here is to double-check that the print job command actually started with an actual print file. It looks like even if the file isn't located, it still starts up the printing process (perhaps as a way to let the client stream commmands to the printer in real-time?). We need to figure out if the printer actually found our file and cancel the print job when it doesn't...